fix highlight behavior
